+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Performance improvements in link/outline remapping
 
+- `dump_files` output now includes attachment 'location'
+
 ## [0.19.0] - 2026-06-01
 
 ### Fixed